#  common: add for_each_substr() for cheap string split