adds a new templated `for_each_substr()` function for string split:
* using `boost::string_view` avoids copies (and potential allocation) of the substrings
* using a callback model avoids baking in the container type, along with the copies and allocations required for insertion
* this interface is ideal for cases where you just want to iterate over the substrings, and don't actually need to store them in a container

example usage:
```c++
const char* input = "a b c d e f";
ceph::for_each_substr(input, " ",
  [] (boost::string_view s) {
    std::cout << "token: " << s << std::endl;
  });
```

by reimplementing `get_str_list()` and `get_str_vec()` in terms of `for_each_substr()`, benchmark results show a speedup of ~30% (where the `Small` variants use tokens small enough for the small string optimization)

before:
```
Benchmark                          Time           CPU Iterations
-----------------------------------------------------------------
BenchStrList/SmallStrList        882 ns        880 ns     749334
BenchStrList/LargeStrList       2015 ns       2012 ns     345552
BenchStrVec/SmallStrVec          600 ns        599 ns    1146149
BenchStrVec/LargeStrVec         1813 ns       1810 ns     381603
```
after:
```
Benchmark                          Time           CPU Iterations    [Speedup]
-----------------------------------------------------------------
BenchStrList/SmallStrList        650 ns        648 ns     929445    1.35
BenchStrList/LargeStrList       1498 ns       1495 ns     441588    1.35
BenchStrVec/SmallStrVec          445 ns        444 ns    1520766    1.34
BenchStrVec/LargeStrVec         1419 ns       1416 ns     462015    1.27
```

(this pr adds a submodule for the google benchmark library, https://github.com/google/benchmark. we don't necessarily need to merge that or the benchmark itself)