@cbodley 2017-11-10
updated. i removed the benchmark stuff, but those commits are still available in https://github.com/cbodley/ceph/commits/wip-str-list-view-bench if anyone wants to reproduce