@tchaikov 2017-11-14
- http://pulpito.ceph.com/kchai-2017-11-13_07:26:34-rados-wip-kefu-testing-2017-11-13-1300-distro-basic-mira/
- http://pulpito.ceph.com/kchai-2017-11-14_05:25:42-rados-wip-kefu-testing-2017-11-14-1048-distro-basic-mira/